added all functionality for service layer
